update-post-controller

+++ utils/saveImage.js
@@ -1,13 +1,16 @@
 import fs from "node:fs";
+import path from "node:path";
+import { v4 as uuidv4 } from "uuid";
 
-export const saveImage = (file, nameIndex, postId) => {
-  const newPath = `./dist/uploads/images/${postId}-image-${nameIndex}.jpg`;
+export const saveImage = (file, postId) => {
+  const imageId = uuidv4();
+  const newPath = `./dist/uploads/images/${postId}-${imageId}.jpg`;
   fs.renameSync(file.path, newPath);
 
   if (process.env.MODE === "development") {
-    return `http://localhost:5000/uploads/images/${postId}-image-${nameIndex}.jpg`;
+    return `http://localhost:5000/uploads/images/${postId}-${imageId}.jpg`;
   } else if (process.env.MODE === "production") {
-    return `https://sigmacuba.com/uploads/images/${postId}-image-${nameIndex}.jpg`;
+    return `https://sigmacuba.com/uploads/images/${postId}-${imageId}.jpg`;
   }
 };
 
@@ -24,8 +27,11 @@ export const saveAvatar = (file, uid) => {
 
 export const updateAvatar = (file, uid) => {
   const path = `./dist/uploads/avatars/${uid}-avatar.jpg`;
-  fs.renameSync(path, `./dist/uploads/avatars/${uid}-old-avatar.jpg`);
-  fs.renameSync(file.path, path);
+
+  if (fs.existsSync(path)) {
+    fs.renameSync(path, `./dist/uploads/avatars/${uid}-old-avatar.jpg`);
+    fs.renameSync(file.path, path);
+  }
 
   if (process.env.MODE === "development") {
     return `http://localhost:5000/uploads/avatars/${uid}-avatar.jpg`;
@@ -48,4 +54,16 @@ export const getStandardAvatarUrl = () => {
   } else if (process.env.MODE === "production") {
     return "https://sigmacuba.com/standard-avatar.jpg";
   }
+};
+
+export const removeImage = (filename) => {
+  const folder = `./dist/uploads/images`;
+  const filePath = path.join(folder, filename);
+
+  if (fs.existsSync(filePath)) {
+    fs.renameSync(filePath, `./dist/uploads/images/old-${filename}`);
+    return true;
+  }
+
+  return false;
 };
